# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gerenciamento de Consultas",
        "description": "API para gerenciamento de consultas médicas",
        "contact": {
            "email": "seu-email@exemplo.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/pacientes": {
            "get": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Lista todos os pacientes",
                "description": "Retorna a lista de pacientes cadastrados no sistema.",
                "operationId": "f198f44bc404f0225a44e06fcb996ac3",
                "responses": {
                    "200": {
                        "description": "Lista de pacientes recuperada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lista de pacientes recuperada com sucesso"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Cadastra um novo paciente",
                "description": "Cria um novo cadastro de paciente no sistema.",
                "operationId": "47eb9d241988aecda161f76fe9047698",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "idade"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Maria Santos"
                                    },
                                    "idade": {
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paciente cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente cadastrado com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/pacientes/{id}": {
            "put": {
                "tags": [
                    "Pacientes"
                ],
                "summary": "Atualiza um paciente",
                "description": "Atualiza os dados de um paciente específico.",
                "operationId": "d3e21cfe04d93e2242aecf60d9756a6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paciente atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Paciente atualizado com sucesso"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}